test(newsletters): split Batch B tests into per-block files
